Location rest api secured

+++ mushroomhunter/mushroomhunter-Web/src/main/java/cz/fi/muni/pa165/mushroomhunter/rest/LocationRest.java
@@ -10,10 +10,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import cz.fi.muni.pa165.mushroomhunter.api.dto.LocationDto;
 import cz.fi.muni.pa165.mushroomhunter.api.service.LocationService;
+import cz.fi.muni.pa165.mushroomhunter.api.service.SecurityService;
 import java.util.ArrayList;
 import java.util.List;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -27,6 +29,9 @@ public class LocationRest {
 
     @Autowired
     LocationService locationService;
+    
+    @Autowired
+    SecurityService securityService;
 
     @RequestMapping(method = RequestMethod.GET)
     public List<LocationDto> getLocationList() {
@@ -53,19 +58,29 @@ public class LocationRest {
 
     @RequestMapping(method = RequestMethod.POST)
     public List<Long> createLocation(@RequestBody @Valid LocationDto location) {
-        List<Long> resultList = new ArrayList<>();
+        List<Long> resultList = new ArrayList<>();      
+        Long currentUserId = securityService.getCurrentlyLoggedUser().getId();
+        location.setOwnerId(currentUserId);
         resultList.add(locationService.save(location));
         return resultList;
     }
 
     @RequestMapping(method = RequestMethod.PUT)
     public LocationDto updateLocation(@RequestBody @Valid LocationDto location) {
+        Long currentUserId = securityService.getCurrentlyLoggedUser().getId();
+        if (!securityService.hasPermissionToModifyEntity(location.getOwnerId())){
+            throw new AccessDeniedException("Access denien: User "+currentUserId+" cannot update location "+location.getId());
+        }
         return locationService.update(location);
     }
 
     @RequestMapping(value = "{locationId}", method = RequestMethod.DELETE)
     public void deleteLocation(@PathVariable Long locationId) {
         LocationDto location = locationService.find(locationId);
+        Long currentUserId = securityService.getCurrentlyLoggedUser().getId();
+        if (!securityService.hasPermissionToModifyEntity(location.getOwnerId())){
+            throw new AccessDeniedException("Access denien: User "+currentUserId+" cannot update location "+location.getId());
+        }
         locationService.delete(location);
     }
 }
